refactoring
I gotta stop doing so much ~and then~ committing.
Primarily, this changes how Map analyzes the board to generate color-maps for a unit. It's more of a readability change than anything else. It's also only partially complete.

Secondly, I did a lot more, too much to keep track of.
- Map and its related classes are now grouped in a folder
- Slider.ts has updated a method name from value() to output(), which I think is clearer.
- Some old, deprecated classes were removed.
- Some preliminary class files were added.
- Fixed a bug I introduced a while back: InfoWindow now updates even when it's on the left side of the screen.
Actually, that might be most of it.
There were a lot of changed files becuase of the refactor.